requirements.txt
flask>=2.2,<4.0
markdown>=3.4,<4.0
psycopg2-binary>=2.9,<3.0
openpyxl>=3.1,<4.0
protobuf>=4.0
opentelemetry-proto>=1.20
